Create ATP fileless
Automated investigation (fileless attack)
This scenario simulates a fileless attack that relies on PowerShell. The simulation triggers Automated investigation, which identifies in-memory artifacts and remediates active processes to address a wider variety of advanced breaches.

Copy the contents of the simulation script and run it using Windows PowerShell on an onboarded device with Windows 10 July 2018 Update Preview or later.